Fix code inspection bugs
Fixes the reported bugs by static analyzers checkstyle, findbugs and javaDocs.
